@@ -20,21 +20,23 @@
     ],
     "homepage": "https://github.com/Riverline/worker-bundle",
     "require":{
-        "php":                          ">=5.3",
+        "php":                          ">=5.6",
         "symfony/console":              "^2.0",
         "symfony/dependency-injection": "^2.0",
         "symfony/config":               "^2.0",
         "symfony/http-kernel":          "^2.0"
     },
     "suggest": {
-        "amazonwebservices/aws-sdk-for-php": "*",
-        "aws/aws-sdk-php":                   "*",
-        "predis/predis":                     "*"
+        "amazonwebservices/aws-sdk-for-php": "^1.0",
+        "aws/aws-sdk-php":                   "^2.0",
+        "predis/predis":                     "^1.0",
+        "stomp-php/stomp-php":               "^4.0"
     },
     "require-dev": {
-        "phpunit/phpunit":                   "^4.8",
-        "aws/aws-sdk-php":                   "^2.8",
-        "predis/predis":                     "^1.1"
+        "phpunit/phpunit":     "^4.8",
+        "aws/aws-sdk-php":     "^2.8",
+        "predis/predis":       "^1.1",
+        "stomp-php/stomp-php": "^4.3"
     },
     "autoload":{
         "psr-4":{
